LayoutTests/http/tests/security/webaudio-render-remote-audio-allowed-crossorigin-redirect-expected.txt

Ensure that audio is rendered when tainted by a remote audio resource when CORS is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS outputArray is not silentArray
PASS successfullyParsed is true

TEST COMPLETE
